dumcrown/requirements.txt
Django>=4.2.3<4.3
psycopg2-binary>2.9.6<2.10
asgiref>=3.7.2,<3.8
channels>=4.0.0,<4.1
channels-redis>=4.1.0<4.2
msgpack>=1.0.5<1.1
redis>=4.6.0<4.7
sqlparse>=0.4.4<0.5
tzdata==2023.3
daphne==4.0.0
django-allauth==0.54.0
celery==5.3
djangorestframework==3.14.0